# Adotados/Adotados/AdotadosAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Achados e Amados",
    "version": "v1"
  },
  "paths": {
    "/api/Adotado": {
      "get": {
        "tags": [
          "Adotado"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AdotadoDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Adotado"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AdotadoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AdotadoDTO"
            }
          }
        }
      }
    },
    "/api/Adotado/{id}": {
      "get": {
        "tags": [
          "Adotado"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AdotadoDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Adotado"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AdotadoDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Adotado"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Animal": {
      "get": {
        "tags": [
          "Animal"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AnimalDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Animal"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AnimalDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AnimalDTO"
            }
          }
        }
      }
    },
    "/api/Animal/{id}": {
      "get": {
        "tags": [
          "Animal"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AnimalDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Animal"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AnimalDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Animal"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Cliente": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ClienteDTO"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClienteDTO"
            }
          }
        }
      }
    },
...
    "/api/Cliente/{id}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ClienteDTO"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cliente"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ClienteDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Cliente"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AdotadoDTO": {
      "type": "object",
      "properties": {
        "idAdotado": {
          "format": "int32",
          "type": "integer"
        },
        "dtaAdocao": {
          "format": "date-time",
          "type": "string"
        },
        "idAnimal": {
          "format": "int32",
          "type": "integer"
        },
        "idCliente": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "AnimalDTO": {
      "type": "object",
      "properties": {
        "idAnimal": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "idade": {
          "type": "string"
        },
        "ehVacinado": {
          "type": "boolean"
        },
        "urlImgAnimal": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        },
        "dtaCadastro": {
          "format": "date-time",
          "type": "string"
        },
        "idTpFuncionario": {
          "format": "int32",
          "type": "integer"
        },
        "idTpAnimal": {
          "format": "int32",
          "type": "integer"
        },
        "sexo": {
          "type": "string"
        },
        "raca": {
          "type": "string"
        },
        "ehCastrado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "ClienteDTO": {
      "type": "object",
      "properties": {
        "idCliente": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "dtaNascimento": {
          "format": "date-time",
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "endereco": {
          "type": "string"
        },
        "nro": {
          "type": "string"
        },
        "bairro": {
          "type": "string"
        },
        "cidade": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "urlImgCliente": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Insira o JWT no formato: Bearer {seu_token}"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}